# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Bot Monitor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotInfo {
  id: number;
  isBot: boolean;
  firstName: string;
  username: string;
  canJoinGroups?: boolean;
  canReadAllGroupMessages?: boolean;
  supportsInlineQueries?: boolean;
  isOnline: boolean;
}

export interface TelegramUpdate {
  updateId: number;
  messageId?: number;
  date: number;
  chatId?: number;
  chatTitle?: string;
  chatType?: string;
  fromId?: number;
  fromName?: string;
  fromUsername?: string;
  text?: string;
  type: string;
}

export interface UpdateList {
  updates: TelegramUpdate[];
  total: number;
}

export type StoredMessageRawData = { [key: string]: unknown };

export interface StoredMessage {
  id: number;
  updateId?: number;
  messageId?: number;
  date: string;
  chatId: string;
  chatTitle?: string;
  chatType: string;
  fromId?: string;
  fromName?: string;
  fromUsername?: string;
  text?: string;
  messageType: string;
  rawData?: StoredMessageRawData;
  createdAt: string;
}

export interface MessageList {
  messages: StoredMessage[];
  total: number;
}

export interface SendMessageBody {
  chatId: string;
  text: string;
}

export interface SendMessageResponse {
  success: boolean;
  messageId?: number;
}

export interface DashboardStats {
  totalMessages: number;
  totalChats: number;
  totalUsers: number;
  todayMessages: number;
  botStatus: string;
  lastUpdateTime?: string;
}

export interface ActivityEvent {
  id: number;
  type: string;
  description: string;
  chatTitle?: string;
  fromName?: string;
  timestamp: string;
}

export interface ActivityList {
  activities: ActivityEvent[];
}

export interface ChatInfo {
  chatId: string;
  title?: string;
  type: string;
  messageCount: number;
  lastMessageAt?: string;
}

export interface ChatList {
  chats: ChatInfo[];
}

export interface MessageCount {
  date: string;
  count: number;
}

export interface MessageCountList {
  counts: MessageCount[];
}

export type GetBotUpdatesParams = {
  limit?: number;
  offset?: number;
};

export type GetMessagesParams = {
  limit?: number;
  offset?: number;
  chatId?: string;
};

export type GetActivityParams = {
  limit?: number;
};

export type GetMessageCountsParams = {
  days?: number;
};