# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Bot Monitor API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns Telegram bot profile and status
 * @summary Get bot information
 */
export const GetBotInfoResponse = zod.object({
  id: zod.number(),
  isBot: zod.boolean(),
  firstName: zod.string(),
  username: zod.string(),
  canJoinGroups: zod.boolean().optional(),
  canReadAllGroupMessages: zod.boolean().optional(),
  supportsInlineQueries: zod.boolean().optional(),
  isOnline: zod.boolean(),
});

/**
 * Returns recent updates/messages received by the bot
 * @summary Get recent bot updates
 */
export const getBotUpdatesQueryLimitDefault = 50;

export const GetBotUpdatesQueryParams = zod.object({
  limit: zod.coerce.number().default(getBotUpdatesQueryLimitDefault),
  offset: zod.coerce.number().optional(),
});

export const GetBotUpdatesResponse = zod.object({
  updates: zod.array(
    zod.object({
      updateId: zod.number(),
      messageId: zod.number().optional(),
      date: zod.number(),
      chatId: zod.number().optional(),
      chatTitle: zod.string().optional(),
      chatType: zod.string().optional(),
      fromId: zod.number().optional(),
      fromName: zod.string().optional(),
      fromUsername: zod.string().optional(),
      text: zod.string().optional(),
      type: zod.string(),
    }),
  ),
  total: zod.number(),
});

/**
 * Returns paginated stored messages
 * @summary Get stored messages
 */
export const getMessagesQueryLimitDefault = 50;
export const getMessagesQueryOffsetDefault = 0;

export const GetMessagesQueryParams = zod.object({
  limit: zod.coerce.number().default(getMessagesQueryLimitDefault),
  offset: zod.coerce.number().default(getMessagesQueryOffsetDefault),
  chatId: zod.coerce.string().optional(),
});

export const GetMessagesResponse = zod.object({
  messages: zod.array(
    zod.object({
      id: zod.number(),
      updateId: zod.number().optional(),
      messageId: zod.number().optional(),
      date: zod.coerce.date(),
      chatId: zod.string(),
      chatTitle: zod.string().optional(),
      chatType: zod.string(),
      fromId: zod.string().optional(),
      fromName: zod.string().optional(),
      fromUsername: zod.string().optional(),
      text: zod.string().optional(),
      messageType: zod.string(),
      rawData: zod.object({}).passthrough().optional(),
      createdAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
});

/**
 * Sends a message to a specific chat
 * @summary Send a message via the bot
 */
export const SendMessageBody = zod.object({
  chatId: zod.string(),
  text: zod.string(),
});

export const SendMessageResponse = zod.object({
  success: zod.boolean(),
  messageId: zod.number().optional(),
});

/**
 * Returns summary stats for the bot
 * @summary Get dashboard statistics
 */
export const GetDashboardStatsResponse = zod.object({
  totalMessages: zod.number(),
  totalChats: zod.number(),
  totalUsers: zod.number(),
  todayMessages: zod.number(),
  botStatus: zod.string(),
  lastUpdateTime: zod.coerce.date().optional(),
});

/**
 * Returns recent activity events
 * @summary Get recent activity timeline
 */
export const getActivityQueryLimitDefault = 20;

export const GetActivityQueryParams = zod.object({
  limit: zod.coerce.number().default(getActivityQueryLimitDefault),
});

export const GetActivityResponse = zod.object({
  activities: zod.array(
    zod.object({
      id: zod.number(),
      type: zod.string(),
      description: zod.string(),
      chatTitle: zod.string().optional(),
      fromName: zod.string().optional(),
      timestamp: zod.coerce.date(),
    }),
  ),
});

/**
 * Returns list of chats the bot is part of
 * @summary Get chat list
 */
export const GetChatsResponse = zod.object({
  chats: zod.array(
    zod.object({
      chatId: zod.string(),
      title: zod.string().optional(),
      type: zod.string(),
      messageCount: zod.number(),
      lastMessageAt: zod.coerce.date().optional(),
    }),
  ),
});

/**
 * Returns message counts grouped by day for charting
 * @summary Get message counts by day
 */
export const getMessageCountsQueryDaysDefault = 7;

export const GetMessageCountsQueryParams = zod.object({
  days: zod.coerce.number().default(getMessageCountsQueryDaysDefault),
});

export const GetMessageCountsResponse = zod.object({
  counts: zod.array(
    zod.object({
      date: zod.string(),
      count: zod.number(),
    }),
  ),
});