# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Bot Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityEvent {
  id: number;
  type: string;
  description: string;
  chatTitle?: string;
  fromName?: string;
  timestamp: Date;
}