# lib/api-zod/src/generated/types/activityList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Bot Monitor API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEvent } from "./activityEvent";

export interface ActivityList {
  activities: ActivityEvent[];
}