# lib/api-zod/src/generated/types/botInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Bot Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface BotInfo {
  id: number;
  isBot: boolean;
  firstName: string;
  username: string;
  canJoinGroups?: boolean;
  canReadAllGroupMessages?: boolean;
  supportsInlineQueries?: boolean;
  isOnline: boolean;
}