# lib/api-zod/src/generated/types/chatInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Bot Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatInfo {
  chatId: string;
  title?: string;
  type: string;
  messageCount: number;
  lastMessageAt?: Date;
}