# lib/api-zod/src/generated/types/messageList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Bot Monitor API
 * OpenAPI spec version: 0.1.0
 */
import type { StoredMessage } from "./storedMessage";

export interface MessageList {
  messages: StoredMessage[];
  total: number;
}