# lib/api-zod/src/generated/types/telegramUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Bot Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramUpdate {
  updateId: number;
  messageId?: number;
  date: number;
  chatId?: number;
  chatTitle?: string;
  chatType?: string;
  fromId?: number;
  fromName?: string;
  fromUsername?: string;
  text?: string;
  type: string;
}